requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# HTTP Requests
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Data Visualization
plotly==5.18.0
dash==2.14.2

# Scheduling
apscheduler==3.10.4

# Date/Time utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# WSGI Server
gunicorn==21.2.0

# TUI (Terminal User Interface)
textual>=0.60.0
rich>=13.0.0
